opsui/traits: relocate use'Trait' composables into traits/'trait' folder and update component imports
Jira: N/A
