+++ ui/conductor/src/traits/occupancy/useOccupancyTrait.js
